refactor merge from private
