+++ planning/behavior_velocity_intersection_module/src/scene_intersection.cpp
@@ -31,7 +31,6 @@
 
 #include <behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
 #include <behavior_velocity_planner_common/utilization/path_utilization.hpp>
-#include <behavior_velocity_planner_common/utilization/trajectory_utils.hpp>
 #include <behavior_velocity_planner_common/utilization/util.hpp>
 #include <magic_enum.hpp>
 #include <opencv2/imgproc.hpp>
@@ -70,259 +69,6 @@ static bool isTargetCollisionVehicleType(
   return false;
 }
 
-static bool isTargetStuckVehicleType(
-  const autoware_auto_perception_msgs::msg::PredictedObject & object)
-{
-  if (
-    object.classification.at(0).label ==
-      autoware_auto_perception_msgs::msg::ObjectClassification::CAR ||
-    object.classification.at(0).label ==
-      autoware_auto_perception_msgs::msg::ObjectClassification::BUS ||
-    object.classification.at(0).label ==
-      autoware_auto_perception_msgs::msg::ObjectClassification::TRUCK ||
-    object.classification.at(0).label ==
-      autoware_auto_perception_msgs::msg::ObjectClassification::TRAILER ||
-    object.classification.at(0).label ==
-      autoware_auto_perception_msgs::msg::ObjectClassification::MOTORCYCLE) {
-    return true;
-  }
-  return false;
-}
-
-static geometry_msgs::msg::Pose getObjectPoseWithVelocityDirection(
-  const autoware_auto_perception_msgs::msg::PredictedObjectKinematics & obj_state)
-{
-  if (obj_state.initial_twist_with_covariance.twist.linear.x >= 0) {
-    return obj_state.initial_pose_with_covariance.pose;
-  }
-
-  // When the object velocity is negative, invert orientation (yaw)
-  auto obj_pose = obj_state.initial_pose_with_covariance.pose;
-  double yaw, pitch, roll;
-  tf2::getEulerYPR(obj_pose.orientation, yaw, pitch, roll);
-  tf2::Quaternion inv_q;
-  inv_q.setRPY(roll, pitch, yaw + M_PI);
-  obj_pose.orientation = tf2::toMsg(inv_q);
-  return obj_pose;
-}
-
-static lanelet::ConstLanelets getEgoLaneWithNextLane(
-  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
-  const std::set<int> & associative_ids, const double width)
-{
-  // NOTE: findLaneIdsInterval returns (start, end) of associative_ids
-  const auto ego_lane_interval_opt = util::findLaneIdsInterval(path, associative_ids);
-  if (!ego_lane_interval_opt) {
-    return lanelet::ConstLanelets({});
-  }
-  const auto [ego_start, ego_end] = ego_lane_interval_opt.value();
-  if (ego_end < path.points.size() - 1) {
-    const int next_id = path.points.at(ego_end).lane_ids.at(0);
-    const auto next_lane_interval_opt = util::findLaneIdsInterval(path, {next_id});
-    if (next_lane_interval_opt) {
-      const auto [next_start, next_end] = next_lane_interval_opt.value();
-      return {
-        planning_utils::generatePathLanelet(path, ego_start, next_start + 1, width),
-        planning_utils::generatePathLanelet(path, next_start + 1, next_end, width)};
-    }
-  }
-  return {planning_utils::generatePathLanelet(path, ego_start, ego_end, width)};
-}
-
-static bool checkStuckVehicleInIntersection(
-  const autoware_auto_perception_msgs::msg::PredictedObjects::ConstSharedPtr objects_ptr,
-  const Polygon2d & stuck_vehicle_detect_area, const double stuck_vehicle_vel_thr,
-  IntersectionModule::DebugData * debug_data)
-{
-  for (const auto & object : objects_ptr->objects) {
-    if (!isTargetStuckVehicleType(object)) {
-      continue;  // not target vehicle type
-    }
-    const auto obj_v = std::fabs(object.kinematics.initial_twist_with_covariance.twist.linear.x);
-    if (obj_v > stuck_vehicle_vel_thr) {
-      continue;  // not stop vehicle
-    }
-
-    // check if the footprint is in the stuck detect area
-    const auto obj_footprint = tier4_autoware_utils::toPolygon2d(object);
-    const bool is_in_stuck_area = !bg::disjoint(obj_footprint, stuck_vehicle_detect_area);
-    if (is_in_stuck_area && debug_data) {
-      debug_data->stuck_targets.objects.push_back(object);
-      return true;
-    }
-  }
-  return false;
-}
-
-static Polygon2d generateStuckVehicleDetectAreaPolygon(
-  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
-  const lanelet::ConstLanelets & ego_lane_with_next_lane, const int closest_idx,
-  const double stuck_vehicle_detect_dist, const double stuck_vehicle_ignore_dist,
-  const double vehicle_length_m)
-{
-  using lanelet::utils::getArcCoordinates;
-  using lanelet::utils::getLaneletLength3d;
-  using lanelet::utils::getPolygonFromArcLength;
-  using lanelet::utils::to2D;
-
-  const double extra_dist = stuck_vehicle_detect_dist + vehicle_length_m;
-  const double ignore_dist = stuck_vehicle_ignore_dist + vehicle_length_m;
-
-  const double intersection_exit_length = getLaneletLength3d(ego_lane_with_next_lane.front());
-
-  const auto closest_arc_coords = getArcCoordinates(
-    ego_lane_with_next_lane, tier4_autoware_utils::getPose(path.points.at(closest_idx).point));
-
-  const double start_arc_length = intersection_exit_length - ignore_dist > closest_arc_coords.length
-                                    ? intersection_exit_length - ignore_dist
-                                    : closest_arc_coords.length;
-
-  const double end_arc_length = getLaneletLength3d(ego_lane_with_next_lane.front()) + extra_dist;
-
-  const auto target_polygon =
-    to2D(getPolygonFromArcLength(ego_lane_with_next_lane, start_arc_length, end_arc_length))
-      .basicPolygon();
-
-  Polygon2d polygon{};
-
-  if (target_polygon.empty()) {
-    return polygon;
-  }
-
-  for (const auto & p : target_polygon) {
-    polygon.outer().emplace_back(p.x(), p.y());
-  }
-
-  polygon.outer().emplace_back(polygon.outer().front());
-  bg::correct(polygon);
-
-  return polygon;
-}
-
-static bool checkAngleForTargetLanelets(
-  const geometry_msgs::msg::Pose & pose, const lanelet::ConstLanelets & target_lanelets,
-  const double detection_area_angle_thr, const double margin = 0.0)
-{
-  for (const auto & ll : target_lanelets) {
-    if (!lanelet::utils::isInLanelet(pose, ll, margin)) {
-      continue;
-    }
-    const double ll_angle = lanelet::utils::getLaneletAngle(ll, pose.position);
-    const double pose_angle = tf2::getYaw(pose.orientation);
-    const double angle_diff = tier4_autoware_utils::normalizeRadian(ll_angle - pose_angle);
-    if (std::fabs(angle_diff) < detection_area_angle_thr) {
-      return true;
-    }
-  }
-  return false;
-}
-
-static void cutPredictPathWithDuration(
-  autoware_auto_perception_msgs::msg::PredictedObjects * objects_ptr,
-  const rclcpp::Clock::SharedPtr clock, const double time_thr)
-{
-  const rclcpp::Time current_time = clock->now();
-  for (auto & object : objects_ptr->objects) {                         // each objects
-    for (auto & predicted_path : object.kinematics.predicted_paths) {  // each predicted paths
-      const auto origin_path = predicted_path;
-      predicted_path.path.clear();
-
-      for (size_t k = 0; k < origin_path.path.size(); ++k) {  // each path points
-        const auto & predicted_pose = origin_path.path.at(k);
-        const auto predicted_time =
-          rclcpp::Time(objects_ptr->header.stamp) +
-          rclcpp::Duration(origin_path.time_step) * static_cast<double>(k);
-        if ((predicted_time - current_time).seconds() < time_thr) {
-          predicted_path.path.push_back(predicted_pose);
-        }
-      }
-    }
-  }
-}
-
-static TimeDistanceArray calcIntersectionPassingTime(
-  const autoware_auto_planning_msgs::msg::PathWithLaneId & path,
-  const std::shared_ptr<const PlannerData> & planner_data, const std::set<int> & associative_ids,
-  const int closest_idx, const double time_delay, const double intersection_velocity,
-  const double minimum_ego_velocity)
-{
-  double dist_sum = 0.0;
-  int assigned_lane_found = false;
-
-  // crop intersection part of the path, and set the reference velocity to intersection_velocity
-  // for ego's ttc
-  PathWithLaneId reference_path;
-  for (size_t i = closest_idx; i < path.points.size(); ++i) {
-    auto reference_point = path.points.at(i);
-    reference_point.point.longitudinal_velocity_mps = intersection_velocity;
-    reference_path.points.push_back(reference_point);
-    bool has_objective_lane_id = util::hasLaneIds(path.points.at(i), associative_ids);
-    if (assigned_lane_found && !has_objective_lane_id) {
-      break;
-    }
-    assigned_lane_found = has_objective_lane_id;
-  }
-  if (!assigned_lane_found) {
-    return {{0.0, 0.0}};  // has already passed the intersection.
-  }
-
-  // apply smoother to reference velocity
-  PathWithLaneId smoothed_reference_path = reference_path;
-  smoothPath(reference_path, smoothed_reference_path, planner_data);
-
-  // calculate when ego is going to reach each (interpolated) points on the path
-  TimeDistanceArray time_distance_array{};
-  dist_sum = 0.0;
-  double passing_time = time_delay;
-  time_distance_array.emplace_back(passing_time, dist_sum);
-  for (size_t i = 1; i < smoothed_reference_path.points.size(); ++i) {
-    const auto & p1 = smoothed_reference_path.points.at(i - 1);
-    const auto & p2 = smoothed_reference_path.points.at(i);
-
-    const double dist = tier4_autoware_utils::calcDistance2d(p1, p2);
-    dist_sum += dist;
-
-    // use average velocity between p1 and p2
-    const double average_velocity =
-      (p1.point.longitudinal_velocity_mps + p2.point.longitudinal_velocity_mps) / 2.0;
-    passing_time +=
-      (dist / std::max<double>(
-                minimum_ego_velocity,
-                average_velocity));  // to avoid zero-division
-
-    time_distance_array.emplace_back(passing_time, dist_sum);
-  }
-
-  return time_distance_array;
-}
-
-static double calcDistanceUntilIntersectionLanelet(
-  const lanelet::ConstLanelet & assigned_lanelet,
-  const autoware_auto_planning_msgs::msg::PathWithLaneId & path, const size_t closest_idx)
-{
-  const auto lane_id = assigned_lanelet.id();
-  const auto intersection_first_itr =
-    std::find_if(path.points.cbegin(), path.points.cend(), [&](const auto & p) {
-      return std::find(p.lane_ids.begin(), p.lane_ids.end(), lane_id) != p.lane_ids.end();
-    });
-  if (
-    intersection_first_itr == path.points.begin() || intersection_first_itr == path.points.end()) {
-    return 0.0;
-  }
-  const auto dst_idx = std::distance(path.points.begin(), intersection_first_itr) - 1;
-
-  if (closest_idx > static_cast<size_t>(dst_idx)) {
-    return 0.0;
-  }
-
-  double distance = std::abs(motion_utils::calcSignedArcLength(path.points, closest_idx, dst_idx));
-  const auto & lane_first_point = assigned_lanelet.centerline2d().front();
-  distance += std::hypot(
-    path.points.at(dst_idx).point.pose.position.x - lane_first_point.x(),
-    path.points.at(dst_idx).point.pose.position.y - lane_first_point.y());
-  return distance;
-}
-
 IntersectionModule::IntersectionModule(
   const int64_t module_id, const int64_t lane_id, std::shared_ptr<const PlannerData> planner_data,
   const PlannerParam & planner_param, const std::set<int> & associative_ids,
@@ -345,7 +91,6 @@ IntersectionModule::IntersectionModule(
   collision_state_machine_.setMarginTime(
     planner_param_.collision_detection.state_transit_margin_time);
   before_creep_state_machine_.setMarginTime(planner_param_.occlusion.before_creep_stop_time);
-  // TODO(Mamoru Sobue): maybe optional is better
   before_creep_state_machine_.setState(StateMachine::State::STOP);
   if (enable_occlusion_detection) {
     occlusion_grid_pub_ = node_.create_publisher<grid_map_msgs::msg::GridMap>(
@@ -511,7 +256,7 @@ void IntersectionModule::prepareRTCStatus(
 
 bool IntersectionModule::modifyPathVelocity(PathWithLaneId * path, StopReason * stop_reason)
 {
-  debug_data_ = DebugData();
+  debug_data_ = util::DebugData();
   *stop_reason = planning_utils::initializeStopReason(StopReason::INTERSECTION);
 
   // set default RTC
@@ -626,8 +371,8 @@ IntersectionModule::DecisionResult IntersectionModule::modifyPathVelocityDetail(
     return IntersectionModule::Indecisive{};
   }
 
-  const auto ego_lane_with_next_lane =
-    getEgoLaneWithNextLane(*path, associative_ids_, planner_data_->vehicle_info_.vehicle_width_m);
+  const auto ego_lane_with_next_lane = util::getEgoLaneWithNextLane(
+    *path, associative_ids_, planner_data_->vehicle_info_.vehicle_width_m);
   const auto check_stuck_vehicle = checkStuckVehicle(
     planner_data_, interpolated_path_info, ego_lane_with_next_lane, first_conflicting_area.value(),
     path);
@@ -811,7 +556,7 @@ std::optional<std::pair<size_t, bool>> IntersectionModule::checkStuckVehicle(
   // considering lane change in the intersection, these lanelets are generated from the path
   const auto ego_lane = ego_lane_with_next_lane.front();
   debug_data_.ego_lane = ego_lane.polygon3d();
-  const auto stuck_vehicle_detect_area = generateStuckVehicleDetectAreaPolygon(
+  const auto stuck_vehicle_detect_area = util::generateStuckVehicleDetectAreaPolygon(
     *input_path, ego_lane_with_next_lane, closest_idx,
     planner_param_.stuck_vehicle.stuck_vehicle_detect_dist,
     planner_param_.stuck_vehicle.stuck_vehicle_ignore_dist,
@@ -837,7 +582,7 @@ std::optional<std::pair<size_t, bool>> IntersectionModule::checkStuckVehicle(
 
   bool is_stuck = false;
   if (!is_over_stuck_stopline) {
-    is_stuck = checkStuckVehicleInIntersection(
+    is_stuck = util::checkStuckVehicleInIntersection(
       objects_ptr, stuck_vehicle_detect_area, planner_param_.stuck_vehicle.stuck_vehicle_vel_thr,
       &debug_data_);
   }
@@ -867,8 +612,8 @@ bool IntersectionModule::checkCollision(
     }
 
     // check direction of objects
-    const auto object_direction = getObjectPoseWithVelocityDirection(object.kinematics);
-    const auto is_in_adjacent_lanelets = checkAngleForTargetLanelets(
+    const auto object_direction = util::getObjectPoseWithVelocityDirection(object.kinematics);
+    const auto is_in_adjacent_lanelets = util::checkAngleForTargetLanelets(
       object_direction, adjacent_lanelets, planner_param_.common.detection_area_angle_thr,
       planner_param_.common.detection_area_margin);
     if (is_in_adjacent_lanelets) {
@@ -881,13 +626,13 @@ bool IntersectionModule::checkCollision(
       const auto is_in_intersection_area = bg::within(obj_poly, intersection_area_2d);
       if (is_in_intersection_area) {
         target_objects.objects.push_back(object);
-      } else if (checkAngleForTargetLanelets(
+      } else if (util::checkAngleForTargetLanelets(
                    object_direction, detection_area_lanelets,
                    planner_param_.common.detection_area_angle_thr,
                    planner_param_.common.detection_area_margin)) {
         target_objects.objects.push_back(object);
       }
-    } else if (checkAngleForTargetLanelets(
+    } else if (util::checkAngleForTargetLanelets(
                  object_direction, detection_area_lanelets,
                  planner_param_.common.detection_area_angle_thr,
                  planner_param_.common.detection_area_margin)) {
@@ -898,18 +643,18 @@ bool IntersectionModule::checkCollision(
 
   // check collision between target_objects predicted path and ego lane
   // cut the predicted path at passing_time
-  const auto time_distance_array = calcIntersectionPassingTime(
+  const auto time_distance_array = util::calcIntersectionPassingTime(
     path, planner_data_, associative_ids_, closest_idx, time_delay,
     planner_param_.common.intersection_velocity,
     planner_param_.collision_detection.minimum_ego_predicted_velocity);
   const double passing_time = time_distance_array.back().first;
-  cutPredictPathWithDuration(&target_objects, clock_, passing_time);
+  util::cutPredictPathWithDuration(&target_objects, clock_, passing_time);
 
   const auto closest_arc_coords = getArcCoordinates(
     ego_lane_with_next_lane, tier4_autoware_utils::getPose(path.points.at(closest_idx).point));
   const auto & ego_lane = ego_lane_with_next_lane.front();
   const double distance_until_intersection =
-    calcDistanceUntilIntersectionLanelet(ego_lane, path, closest_idx);
+    util::calcDistanceUntilIntersectionLanelet(ego_lane, path, closest_idx);
   const double base_link2front = planner_data_->vehicle_info_.max_longitudinal_offset_m;
 
   const auto ego_poly = ego_lane.polygon2d().basicPolygon();
